+++ mayan/apps/document_exports/tests/mixins.py
@@ -1,4 +1,7 @@
-class DocumentVersionExportAPIViewTestMixin:
+from mayan.apps.documents.tests.mixins.document_mixins import DocumentTestMixin
+
+
+class DocumentVersionExportAPIViewTestMixin(DocumentTestMixin):
     def _request_test_document_version_export_api_view_via_get(self):
         return self.get(
             viewname='rest_api:documentversion-export', kwargs={
@@ -16,7 +19,7 @@ class DocumentVersionExportAPIViewTestMixin:
         )
 
 
-class DocumentVersionExportViewTestMixin:
+class DocumentVersionExportViewTestMixin(DocumentTestMixin):
     def _request_test_document_version_export_view(self):
         return self.post(
             viewname='document_exports:document_version_export', kwargs={