+++ mayan/apps/documents/links/document_file_links.py
@@ -7,16 +7,16 @@ from mayan.apps.navigation.classes import Link
 
 from ..icons import (
     icon_document_file_delete, icon_document_file_edit,
-    icon_document_file_list, icon_document_file_print,
-    icon_document_file_properties_detail,
+    icon_document_file_introspect, icon_document_file_list,
+    icon_document_file_print, icon_document_file_properties_detail,
     icon_document_file_return_to_document, icon_document_file_return_list,
     icon_document_file_preview, icon_document_file_transformation_list_clear,
     icon_document_file_transformation_list_clone
 )
 from ..permissions import (
     permission_document_file_delete, permission_document_file_edit,
-    permission_document_file_print, permission_document_file_view,
-    permission_document_view
+    permission_document_file_print, permission_document_file_tools,
+    permission_document_file_view, permission_document_view
 )
 
 link_document_file_delete = Link(
@@ -34,6 +34,18 @@ link_document_file_edit = Link(
     args='object.pk', icon=icon_document_file_edit,
     permissions=(permission_document_file_edit,),
     text=_('Edit'), view='documents:document_file_edit',
+)
+link_document_file_introspect_multiple = Link(
+    icon=icon_document_file_introspect,
+    text=_('Introspect'),
+    view='documents:document_file_introspect_multiple'
+)
+link_document_file_introspect_single = Link(
+    args='resolved_object.pk',
+    icon=icon_document_file_introspect,
+    permissions=(permission_document_file_tools,),
+    text=_('Introspect'),
+    view='documents:document_file_introspect_single'
 )
 link_document_file_list = Link(
     args='resolved_object.pk',