+++ mayan/apps/documents/links/document_file_page_links.py
@@ -13,12 +13,10 @@ from ..icons import (
     icon_document_file_page_rotate_left,
     icon_document_file_page_rotate_right, icon_document_file_page_zoom_in,
     icon_document_file_page_zoom_out, icon_document_file_page_detail,
-    icon_document_file_page_detail_reset, icon_document_file_page_count_update,
-    icon_document_file_page_list
+    icon_document_file_page_detail_reset, icon_document_file_page_list
 )
 from ..permissions import (
-    permission_document_file_tools, permission_document_file_view,
-    permission_document_view
+    permission_document_file_view, permission_document_view
 )
 from ..settings import setting_zoom_max_level, setting_zoom_min_level
 
@@ -121,16 +119,4 @@ link_document_file_page_list = Link(
     args='resolved_object.pk', icon=icon_document_file_page_list,
     permissions=(permission_document_file_view,), text=_('Pages'),
     view='documents:document_file_page_list'
-)
-link_document_file_multiple_page_count_update = Link(
-    icon=icon_document_file_page_count_update,
-    text=_('Recalculate page count'),
-    view='documents:document_file_multiple_page_count_update'
-)
-link_document_file_page_count_update = Link(
-    args='resolved_object.pk',
-    icon=icon_document_file_page_count_update,
-    permissions=(permission_document_file_tools,),
-    text=_('Recalculate page count'),
-    view='documents:document_file_page_count_update'
 )