+++ mayan/apps/documents/tests/mixins/document_mixins.py
@@ -109,12 +109,12 @@ class DocumentTestMixin(DocumentTypeTestMixin):
         Layer.invalidate_cache()
 
         self._test_documents = []
-        self._test_document_files = []
-        self._test_document_file_pages = []
+        self._test_document_file_list = []
+        self._test_document_file_page_list = []
         self._test_document_id_list = []
         self._test_document_id_list_string = []
-        self._test_document_versions = []
-        self._test_document_version_pages = []
+        self._test_document_version_list = []
+        self._test_document_version_page_list = []
 
         if self.auto_create_test_document_type:
             if self._test_document_count > 1:
@@ -203,14 +203,14 @@ class DocumentTestMixin(DocumentTypeTestMixin):
         )
 
         self._test_document_file = document.file_latest
-        self._test_document_files.append(self._test_document_file)
-        self._test_document_file_pages = list(
+        self._test_document_file_list.append(self._test_document_file)
+        self._test_document_file_page_list = list(
             self._test_document_file.file_pages.all()
         )
         self._test_document_file_page = self._test_document_file.file_pages.first()
         self._test_document_version = self._test_document.version_active
-        self._test_document_versions.append(self._test_document_version)
-        self._test_document_version_pages = list(
+        self._test_document_version_list.append(self._test_document_version)
+        self._test_document_version_page_list = list(
             self._test_document_version.version_pages.all()
         )
         self._test_document_version_page = self._test_document_version.version_pages.first()