+++ mayan/apps/documents/tests/test_document_file_api.py
@@ -5,7 +5,7 @@ from mayan.apps.rest_api.tests.base import BaseAPITestCase
 from ..events import (
     event_document_file_created, event_document_file_deleted,
     event_document_file_edited, event_document_version_created,
-    event_document_version_page_created
+    event_document_version_page_created, event_document_version_page_deleted
 )
 from ..permissions import (
     permission_document_file_delete, permission_document_file_new,
@@ -62,12 +62,19 @@ class DocumentFileAPIViewTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 1)
+        self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
         self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self._test_document)
-        self.assertEqual(events[0].verb, event_document_file_deleted.id)
+        self.assertEqual(events[0].target, self._test_document_version)
+        self.assertEqual(
+            events[0].verb, event_document_version_page_deleted.id
+        )
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, self._test_document)
+        self.assertEqual(events[1].verb, event_document_file_deleted.id)
 
     def test_trashed_document_file_delete_api_view_with_access(self):
         self._upload_test_document()